Add property tests for link text and URL parsing (#386)
Address review feedback on the balanced link parser by adding
property-based coverage for `parse_link_text` and `parse_link_url`.
The new proptest cases exercise arbitrary odd and even backslash runs
and nested parentheses and brackets, asserting that:

- escaped delimiters (behind odd backslash runs) never terminate a span,
- the first subsequent unescaped delimiter does terminate it, and
- balanced nested destinations close at the outer delimiter that returns
  the depth to zero.

The strategies construct only valid inputs from each seed, so the
expected terminator position is known by construction rather than via a
re-implemented oracle. Each property was mutation-checked against a
deliberately broken escape guard to confirm it fails as intended.

To keep `parsing.rs` within the 400-line file limit, the inline test
module is extracted into a sibling `parsing_tests.rs`, wired via
`#[path]` in line with the existing tokenizer test-module convention.
