+++ src/wrap/tokenize/parsing.rs
@@ -219,182 +219,5 @@ pub(super) fn handle_backtick_fence(text: &str, start_idx: usize) -> (String, us
 }
 
 #[cfg(test)]
-mod tests {
-    use proptest::prelude::*;
-    use rstest::rstest;
-
-    use super::*;
-
-    fn footnote_label_part_strategy() -> impl Strategy<Value = String> {
-        prop::collection::vec(
-            prop_oneof![
-                (b'a'..=b'z').prop_map(char::from),
-                (b'A'..=b'Z').prop_map(char::from),
-                (b'0'..=b'9').prop_map(char::from),
-                Just('-'),
-                Just('_')
-            ],
-            0..12,
-        )
-        .prop_map(|chars| chars.into_iter().collect())
-    }
-
-    #[test]
-    fn parse_link_or_image_handles_nested_parentheses() {
-        let text = "![alt](path(a(b)c)) more";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "![alt](path(a(b)c))");
-        assert_eq!(idx, token.len());
-    }
-    #[test]
-    fn parse_link_or_image_falls_back_on_malformed_input() {
-        let text = "[broken";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "[");
-        assert_eq!(idx, "[".len());
-    }
-    #[test]
-    fn parse_link_or_image_handles_deeply_nested_parentheses() {
-        let text = "[link](url(a(b(c)d)e)) tail";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "[link](url(a(b(c)d)e))");
-        assert_eq!(idx, token.len());
-    }
-
-    #[test]
-    fn parse_link_or_image_handles_nested_parentheses_for_images() {
-        let text = "![alt](path(a(b(c)d)e))";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "![alt](path(a(b(c)d)e))");
-        assert_eq!(idx, token.len());
-    }
-
-    #[test]
-    fn parse_link_or_image_handles_text_ending_at_bracket() {
-        let text = "[";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "[");
-        assert_eq!(idx, 1);
-    }
-
-    #[test]
-    fn parse_link_or_image_preserves_footnote_reference() {
-        let text = "[^4] tail";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "[^4]");
-        assert_eq!(idx, token.len());
-    }
-
-    #[test]
-    fn parse_link_or_image_preserves_footnote_reference_with_escaped_bracket() {
-        let text = r"[^a\]b] tail";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, r"[^a\]b]");
-        assert_eq!(idx, token.len());
-    }
-
-    #[test]
-    fn parse_link_or_image_preserves_footnote_at_end() {
-        let text = "[^4]";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "[^4]");
-        assert_eq!(idx, token.len());
-    }
-
-    #[test]
-    fn parse_link_or_image_keeps_caret_text_links_as_links() {
-        let text = "[^label](https://example.com) tail";
-        let (token, idx) = parse_link_or_image(text, 0);
-        assert_eq!(token, "[^label](https://example.com)");
-        assert_eq!(idx, token.len());
-    }
-
-    #[rstest]
-    fn parse_link_or_image_preserves_reference_style_link() {
-        let input = "[trybuild][implicit-fixture-trybuild]";
-
-        assert_eq!(
-            parse_link_or_image(input, 0),
-            (input.to_string(), input.len())
-        );
-    }
-
-    proptest! {
-        #[test]
-        fn parse_link_or_image_preserves_footnote_references_with_escaped_brackets(
-            prefix in footnote_label_part_strategy(),
-            suffix in footnote_label_part_strategy(),
-        ) {
-            let expected = format!(r"[^{prefix}\]{suffix}]");
-            let expected_len = expected.len();
-            let text = format!("{expected} tail");
-
-            let (token, idx) = parse_link_or_image(&text, 0);
-
-            prop_assert_eq!(token, expected);
-            prop_assert_eq!(idx, expected_len);
-        }
-    }
-
-    mod tracing_tests {
-        //! Traced-event tests for the parsing helpers.
-        //!
-        //! Verifies that `parse_link_or_image` and `find_footnote_end` emit
-        //! the expected DEBUG and TRACE events, including structured fields,
-        //! for all reachable branches: footnote reference parsed, link or
-        //! image parsed, prefix mismatch, footnote label span recognized, and
-        //! unterminated bracket.
-
-        use tracing_test::traced_test;
-
-        use super::*;
-
-        #[traced_test]
-        #[test]
-        fn parse_link_or_image_logs_footnote_reference() {
-            let _ = parse_link_or_image("[^4] tail", 0);
-            assert!(logs_contain("footnote reference parsed"));
-            assert!(logs_contain("token_length=4"));
-            assert!(!logs_contain("[^4]"));
-        }
-
-        #[traced_test]
-        #[test]
-        fn parse_link_or_image_logs_link_parsed() {
-            let _ = parse_link_or_image("[link](url)", 0);
-            assert!(logs_contain("link or image parsed"));
-            assert!(logs_contain("token_length=11"));
-            assert!(!logs_contain("[link](url)"));
-            assert!(logs_contain("is_image="));
-        }
-
-        #[traced_test]
-        #[test]
-        fn find_footnote_end_logs_prefix_mismatch() {
-            let _ = find_footnote_end("no-caret", 0);
-            assert!(logs_contain("footnote end not found"));
-            assert!(logs_contain("reason="));
-            assert!(logs_contain("prefix_mismatch"));
-        }
-
-        #[traced_test]
-        #[test]
-        fn parse_link_or_image_logs_footnote_label_span() {
-            let _ = parse_link_or_image("[^4] tail", 0);
-            assert!(logs_contain("footnote label span recognized"));
-            assert!(logs_contain("start="));
-            assert!(logs_contain("end="));
-            assert!(logs_contain("token_length=4"));
-            assert!(!logs_contain("[^4]"));
-        }
-
-        #[traced_test]
-        #[test]
-        fn find_footnote_end_logs_unterminated_bracket() {
-            let _ = find_footnote_end("[^unterminated", 0);
-            assert!(logs_contain("footnote end not found"));
-            assert!(logs_contain("reason="));
-            assert!(logs_contain("unterminated_bracket"));
-        }
-    }
-}
+#[path = "parsing_tests.rs"]
+mod tests;
